% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{SF.RAW.all}
\alias{SF.RAW.all}
\title{Returns a dataset containing raw size-frequency data for all categories of species}
\usage{
SF.RAW.all()
}
\description{
Returns a dataset containing raw size-frequency data for all categories of species
}
